# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ambatuwork API v1",
        "description": "API documentation for Ambatuwork, a Scrum-based collaboration application used by Web and Android clients. Authenticate your requests using a Sanctum Bearer token.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 Endpoint Base"
        }
    ],
    "paths": {
        "/auth/google": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Exchange Google ID Token for API Token",
                "description": "Exchanges a Google ID token for a Sanctum API token. If the user doesn't exist, it creates a new user.",
                "operationId": "cf0eb142a9e03738e824ef19df479b5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_token"
                                ],
                                "properties": {
                                    "id_token": {
                                        "description": "Google ID Token",
                                        "type": "string",
                                        "example": "google-id-token-xyz"
                                    },
                                    "device_name": {
                                        "description": "Device/client identifier",
                                        "type": "string",
                                        "example": "android_client",
                                        "default": "android"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|sanctum-token-example"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content / Validation failure"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get Authenticated User",
                "description": "Returns the currently logged in user's details.",
                "operationId": "e339e0e74860c0647fd1d41bbbbc13b2",
                "responses": {
                    "200": {
                        "description": "Authenticated User Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout Authenticated User",
                "description": "Revokes and deletes the current Sanctum token.",
                "operationId": "a05022a34eef599e04fb0a5254375b84",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}/sprints/{sprint}/checkins": {
            "get": {
                "tags": [
                    "Daily Check-ins"
                ],
                "summary": "List Daily Check-ins",
                "description": "Returns all daily check-ins submitted during the specified sprint.",
                "operationId": "7ebabe52f1bf9225a28e77e97b1e9cc5",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DailyCheckin"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Sprint not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Daily Check-ins"
                ],
                "summary": "Submit Daily Check-in",
                "description": "Submits a member's daily check-in. If the check-in contains blockers, an Impediment is automatically created.",
                "operationId": "4c8e0252b15dd69c7293e23ab44a5a67",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "confidence_score",
                                    "checkin_date"
                                ],
                                "properties": {
                                    "yesterday": {
                                        "type": "string",
                                        "example": "Implemented tests",
                                        "nullable": true
                                    },
                                    "today": {
                                        "type": "string",
                                        "example": "Debugging routes",
                                        "nullable": true
                                    },
                                    "blockers": {
                                        "type": "string",
                                        "example": "Database connection timeout",
                                        "nullable": true
                                    },
                                    "confidence_score": {
                                        "type": "integer",
                                        "example": 4,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "checkin_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-04"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Check-in submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DailyCheckin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Sprint not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "avatar_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://lh3.googleusercontent.com/...",
                        "nullable": true
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-01T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "DailyCheckin": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sprint_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "yesterday": {
                        "type": "string",
                        "example": "Implemented tests",
                        "nullable": true
                    },
                    "today": {
                        "type": "string",
                        "example": "Debugging routes",
                        "nullable": true
                    },
                    "blockers": {
                        "type": "string",
                        "example": "Database connection timeout",
                        "nullable": true
                    },
                    "confidence_score": {
                        "type": "integer",
                        "example": 4
                    },
                    "checkin_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-04"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token. Example: Bearer 1|abcdef...",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}